Getting sanity tests mostly working.
Add avmplus executables so we can run the sanity tests without having to grab a tamarin build from mozilla.  
The avmplus executables are in bin/<windows|mac|linux> directories.

Added playerglobal.abc, builtin.abc, and toplevel.abc to asc/abc directory as those are necessary for running the tests as well.

Cleaned up some of the expected output of the sanities.  The tests were working ok, but there were minor differences in the expected and actual output. 
Removed some tests that ASC doesn't support at the moment.  I moved these tests to asc/test/sanity/unused so if we turn those features back on, we still have the tests for them.

The sanities now mostly pass.  There are still 2 failures which need investigation and may be real issues.


git-svn-id: http://opensource.adobe.com/svn/opensource/flex/sdk/trunk/modules/asc@683 a9308255-753e-0410-a2e9-80b3fbc4fff6