compiler: added a new token {targetPlayerMajorVersion} to support multiple versions of playerglobal.swc * playerglobal.swc has been moved into libs/player/9/playerglobal.swc * {tpmv} is available in all path-elements, e.g. external-library-path, in flex-config, commandline, and in <source-path> Ant task   - it is replaced at compile-time with the major version from --target-player, e.g. 9 * added an OEM API for code model to get namespaces that should be automatically opened wrt target-player. * build scripts reference playerglobal using a new ant property called ${playerglobal.version} * default target-player is 9.0.115
Reviewer: Pete and Paul
Bugs: n/a
QA: Yes, we should test all the targets in build_framework.xml and all other build.xmls that were modified.
Doc: Yes, we need to document the new config token in flex-config and path-elements. This will be back-merged to 3.0.x as well, so those docs probably need to be updated too.


git-svn-id: http://opensource.adobe.com/svn/opensource/flex/sdk/trunk/modules/asc@1448 a9308255-753e-0410-a2e9-80b3fbc4fff6