Fix bugs ASC-3310, ASC-3337
ASC-3337:
ASC was emitting multinames for [] access with the wrong namespace set.  It was using the def_namespaces, which is a list of open namespaces that conflicting definitions could be found in (public, private, protected, internal for a class).  This set does not include any other open namespaces.  This changes asc to use the set of all open namespaces, which includes namespaces open by default, and explicitly by the user.

ASC-3310:
LintEvaluator was returning ObjectValue.undefinedValue sometimes from a method that was supposed to be returning a TypeValue.  This caused a ClaassCastException later as the return value was cast to a TypeValue.  Change to return noType (*) instead of the undefinedValue.

smokes: asc tests, tamarin tests, flex checkintests



git-svn-id: http://opensource.adobe.com/svn/opensource/flex/sdk/trunk/modules/asc@1571 a9308255-753e-0410-a2e9-80b3fbc4fff6