Some minor changes:
integrate some changes to macromedia/abc/Optimizer from the captivate team - can now specify an output file name, and it closes the file after its done writing it.

Add an avmtarget switch to ScriptCompiler

Fix a minor issue with '--' in GlobalOptimizer.java it was treating it as a file and trying to open it, which failed.


git-svn-id: http://opensource.adobe.com/svn/opensource/flex/sdk/trunk/modules/asc@2018 a9308255-753e-0410-a2e9-80b3fbc4fff6