Fix a bugs asc-3384 & asc-3414.
LintEvaluator was just doing the wrong thing for a WithStatement.  Changed it to do what all the other evaluators do.  This caused the reference inside the with to not know it was inside a with statement, and early bind to the local var.  

Changed code in CE to return the ReferenceValue for an ApplyTypeExpr - this was preventing the as operator from working correctly.

flex checkintests, tamarin tests pass.


git-svn-id: http://opensource.adobe.com/svn/opensource/flex/sdk/trunk/modules/asc@2021 a9308255-753e-0410-a2e9-80b3fbc4fff6