merge change from private diesel branch with svn merge -c 2339 http://opensource.adobe.com/svn/opensource/private/flex/sdk/branches/diesel/modules/asc
A bug was injected into the trunk asc by revision 2234, which was
merged into private diesel trunk in revision 2239.  The issue
seems to be that -1 could be written into the ABC as the debug
info line number, which causes VerifyErrors and other bad stuff,
at least in AIR. The fix is just to ouptut 0 as the line number
if it is less than 0.


git-svn-id: http://opensource.adobe.com/svn/opensource/flex/sdk/trunk/modules/asc@2362 a9308255-753e-0410-a2e9-80b3fbc4fff6