Simplified asc Scanner
Removed a lot of code from the asc scan phase in preperation for performance optimization work.
Since a number of methods have been redefined as static, unknown users of these methods should check for and report
new failures soon.

reviewer: Jeff Dyer
tests: checkintests asc-tests

Note that as3/LocalizedErrors/1095-LineTerminatorInDoubleQuotes.as in asc-tests fails due to
a mismatch on the error message produced. This is expected and the asc-tests should be updated.




git-svn-id: http://opensource.adobe.com/svn/opensource/flex/sdk/trunk/modules/asc@2627 a9308255-753e-0410-a2e9-80b3fbc4fff6