Fix asdoc to work somewhat sensibly.
ASDoc comments will now always be associated with the following definition, unless there is intervening metadata that takes a comment.  Currently this is limited to "Style", "Event", and "Effect" metadata.  If there is intervening metadata that is not "Style" "Event" or "Effect" then the asdoc comment will apply to the definition.  This should fix all the problems where other metadata is inserted between the comment and the definition, and suddenly no more documentation is showing up for that definition.

For example, this comment now applies to the class, instead of just disappearing because of the RandomMetadata:

/** My ASDoc Comment, which applies to the class below, not the random metadata in between */
[RandomMetadata]
public class C {...}

Also fixed asdoc so that there is now only 1 entry for each definition in the generated XML.  Before you would sometimes get 1, sometimes 2, sometimes more, and it just confused tools down the line.  Now, the last ASDoc comment is the one output into the XML.

smokes: tamarin tests, flex checkintests, asdoc builds

git-svn-id: http://opensource.adobe.com/svn/opensource/flex/sdk/trunk/modules/asc@4071 a9308255-753e-0410-a2e9-80b3fbc4fff6