the AS3 debugger wasn't stopping at breakpoints in frame scripts that contain include statements
we use the qualified error origin (in Context) for frame script module
name mapping, and when the CodeGenerator processes IncludeDirectiveNodes,
it resets the Emitter's origin, but was using Context.getErrorOrigin
instead of Context.getQualifiedErrorOrigin. it's safe to add this to
core asc (for all clients) because getQualifiedErrorOrigin just passes
through to getErrorOrigin if a qualified one isn't set (is null or
blank).

QE Notes: none
Bugs: authoring bugbase 198885
Reviewer: Erik Tierney



git-svn-id: http://opensource.adobe.com/svn/opensource/flex/sdk/trunk/modules/asc@4082 a9308255-753e-0410-a2e9-80b3fbc4fff6