Adding sample eclipse projects for the Java modules so that they do not rely on a FLEX_SDK linked resource or classpath variable.
The "antTasks" project, however, needs to know the install location of the ANT 1.7.0 lib/ant.jar - so if a developer wants to import this optional project they must still  setup a Java classpath variable "ANT17_JAR" and point it to the ant.jar file. 

A compiler developer should first import the /lib project so that common jars will be found by all of the other projects. After importing this project, the other modules projects can be imported.

No code changes.

QE notes: N/A
Doc notes: N/A
Bugs: N/A
Reviewer: N/A 
Tests run: checkintests
Is noteworthy for integration: No


git-svn-id: http://opensource.adobe.com/svn/opensource/flex/sdk/trunk/modules/asc@14020 a9308255-753e-0410-a2e9-80b3fbc4fff6