Merge 2 changes to trunk so that the ASC that Tamarin uses is up to date:
------------------------------------------------------------------------
r17787 | pfarland@adobe.com | 2010-09-21 14:18:37 -0400 (Tue, 21 Sep 2010) | 9 lines

Merging 4.x revision 17307 to hero mainline to make sure we don't lose these AIR specific updates to ASC for runtime API versioning.

QA notes: N/A
Doc notes: N/A
Bugs: N/A
Reviewer: Frank S-G
Tests run: Checkintests 
Is noteworthy for integration: No

------------------------------------------------------------------------
r17712 | tierney@adobe.com | 2010-09-15 12:50:00 -0400 (Wed, 15 Sep 2010) | 10 lines

Fix ASC bug where the code for named, nested functions was incorrect - the bug was that the function could not recursively call itself, as the code to make the function visible to itself was not being emitted.  This was broken by an optimization to null out the FunctionBuilder for a method after we codegen'ed the method - but, the FunctionBuilder contained data that was needed after it was nulled out.  Moved the flag from the FunctionBuilder to the FunctionCommonNode so we know we always have it.    

QE notes:
Doc notes:
Bugs: ASC-4136
Reviewer: Paul Reilly, Tom Harwood
Tests run: checkintests, full tamarin tests
Is noteworthy for integration: no

git-svn-id: http://opensource.adobe.com/svn/opensource/flex/sdk/trunk/modules/asc@17801 a9308255-753e-0410-a2e9-80b3fbc4fff6